[TASK] Code cleanup based on Psalm static analyser
Code cleanup based on the static analyser Psalm.
https://github.com/vimeo/psalm

Resolves: #79815
Releases: master
Reviewed-on: https://review.typo3.org/51679
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>

+++ typo3/sysext/rtehtmlarea/Configuration/TCA/tx_rtehtmlarea_acronym.php
@@ -44,7 +44,7 @@ return [
                 'default' => 0,
                 'range' => [
                     'upper' => mktime(0, 0, 0, 12, 31, 2020),
-                    'lower' => mktime(0, 0, 0, date('m') - 1, date('d'), date('Y'))
+                    'lower' => mktime(0, 0, 0, (int)date('m') - 1, (int)date('d'), (int)date('Y'))
                 ]
             ]
         ],
